Autoconverted sources with fixes ydanila/sharpen_imazen@836e32ba830ccb31b8ad800d5c151723432d6a73 and updated configuration ydanila/j-metadata-extractor@1be94dcd5965f9e5ccc71b512506303e273cc214

+++ Com.Drew/Com/Drew/Metadata/Exif/ExifReader.cs
@@ -251,7 +251,7 @@ namespace Com.Drew.Metadata.Exif
 				{
 					try
 					{
-						sbyte[] thumbnailData = reader.GetBytes(tiffHeaderOffset + offset, length);
+						sbyte[] thumbnailData = reader.GetBytes(tiffHeaderOffset + (int)offset, (int)length);
 						thumbnailDirectory.SetThumbnailData(thumbnailData);
 					}
 					catch (IOException ex)
